Changes the signature of rectangle from int to double

+++ test/com/bootcamp/tw/p1/shapes/Rectangle.java
@@ -1,15 +1,15 @@
 package com.bootcamp.tw.p1.shapes;
 
 public class Rectangle {
-    private final int length;
-    private final int width;
+    private final double length;
+    private final double width;
 
-    public Rectangle(int length, int width) {
+    public Rectangle(double length, double width) {
         this.length = length;
         this.width = width;
     }
 
-    public int calculateArea() {
+    public double calculateArea() {
         return length * width;
     }
 }
